moved all amount tests to individual file, renamed regression.test.js => runAmount.test.js
